+++ src/main/java/com/gdsc/knu/controller/MarineLifeController.java
@@ -30,7 +30,7 @@ public class MarineLifeController {
     // 사용자가 모은 해양 생물 정보 조회
     @GetMapping("/retrieve/{userID}")
     @Operation(summary = "사용자가 모은 해양 생물 정보 조회", description = "사용자가 모은 해양 생물 정보를 조회한다")
-    public ResponseEntity<?> getMarineLifeByUser(@PathVariable Integer userID) {
+    public ResponseEntity<?> getMarineLifeByUser(@PathVariable Long userID) {
         List<MarineLife> marineLifeList = marineLifeRepository.findByUserId(userID);
         if (marineLifeList.isEmpty()) {
             return new ResponseEntity<>("해당 사용자가 모은 해양 생물 정보가 없습니다.", HttpStatus.NOT_FOUND);
@@ -41,7 +41,7 @@ public class MarineLifeController {
     // 해양 생물 정보 수정
     @PutMapping("/modify/{userID}/{marineLifeID}")
     @Operation(summary = "사용자가 해양 생물 정보 수정하는 기능", description = "사용자가 모은 해양 생물 정보를 수정한다")
-    public ResponseEntity<?> updateMarineLife(@PathVariable Integer userID, @PathVariable Integer marineLifeID, @RequestBody MarineLife newMarineLife) {
+    public ResponseEntity<?> updateMarineLife(@PathVariable Long userID, @PathVariable Long marineLifeID, @RequestBody MarineLife newMarineLife) {
         Optional<MarineLife> optionalMarineLife = marineLifeRepository.findById(marineLifeID);
         if (!optionalMarineLife.isPresent()) {
             return new ResponseEntity<>("해양 생물 정보를 찾을 수 없습니다.", HttpStatus.NOT_FOUND);
@@ -61,7 +61,7 @@ public class MarineLifeController {
     // 해양 생물 정보 삭제
     @DeleteMapping("/delete/{userID}")
     @Operation(summary = "사용자가 모은 해양 생물 정보 삭제", description = "사용자가 생각하기에 이상한 것이 해양 생물로 들어갔다면, 해양 생물 정보를 삭제한다")
-    public ResponseEntity<?> deleteMarineLife(@PathVariable Integer userID, @PathVariable Integer marineLifeID) {
+    public ResponseEntity<?> deleteMarineLife(@PathVariable Long userID, @PathVariable Long marineLifeID) {
         Optional<MarineLife> optionalMarineLife = marineLifeRepository.findById(marineLifeID);
         if (!optionalMarineLife.isPresent()) {
             return new ResponseEntity<>("해양 생물 정보를 찾을 수 없습니다.", HttpStatus.NOT_FOUND);
