+++ src/main/java/com/gdsc/knu/dto/FriendDTO.java
@@ -1,22 +1,22 @@
 package com.gdsc.knu.dto;
 
 public class FriendDTO {
-    private Integer userId;
-    private Integer friendUserId;
+    private Long userId;
+    private Long friendUserId;
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 
-    public Integer getFriendUserId() {
+    public Long getFriendUserId() {
         return friendUserId;
     }
 
-    public void setFriendUserId(Integer friendUserId) {
+    public void setFriendUserId(Long friendUserId) {
         this.friendUserId = friendUserId;
     }
